+++ device-minion-runtime/src/MasterMinion/src/workers/sqw.c
@@ -69,15 +69,15 @@ typedef struct sqw_cmds_status_ {
     \brief Submission Queue Worker Control Block structure
 */
 typedef struct sqw_cb_ {
-    sqw_cmds_status_t   sqw_status[MM_SQ_COUNT];
-    local_fcc_flag_t    sqw_fcc_flags[MM_SQ_COUNT];
+    sqw_cmds_status_t sqw_status[MM_SQ_COUNT];
+    local_fcc_flag_t sqw_fcc_flags[MM_SQ_COUNT];
 } sqw_cb_t;
 
 /*! \var sqw_cb_t SQW_CB
     \brief Global Submission Queue Worker Control Block
     \warning Not thread safe!
 */
-static sqw_cb_t SQW_CB __attribute__((aligned(64))) = {0};
+static sqw_cb_t SQW_CB __attribute__((aligned(64))) = { 0 };
 
 /************************************************************************
 *
@@ -186,8 +186,7 @@ void SQW_Notify(uint8_t sqw_idx)
 
     /* TODO: Future improvements: 1. To use IPIs.
     2. Improve FCC to address security concerns. */
-    local_fcc_flag_notify_no_ack(&SQW_CB.sqw_fcc_flags[sqw_idx],
-        minion, SQW_THREAD_ID);
+    local_fcc_flag_notify_no_ack(&SQW_CB.sqw_fcc_flags[sqw_idx], minion, SQW_THREAD_ID);
 
     return;
 }
@@ -218,13 +217,13 @@ void SQW_Notify(uint8_t sqw_idx)
 *
 ***********************************************************************/
 static inline void sqw_process_waiting_commands(uint32_t sqw_idx, vq_cb_t *vq_cached,
-    vq_cb_t *vq_shared, uint64_t start_cycles, void* shared_mem_ptr, uint64_t vq_used_space)
+    vq_cb_t *vq_shared, uint64_t start_cycles, void *shared_mem_ptr, uint64_t vq_used_space)
 {
-    uint8_t *cmd_buff = (uint8_t*)(MM_SQ_PREFETCHED_BUFFER_BASEADDR + sqw_idx * MM_SQ_SIZE_MAX);
+    uint8_t *cmd_buff = (uint8_t *)(MM_SQ_PREFETCHED_BUFFER_BASEADDR + sqw_idx * MM_SQ_SIZE_MAX);
     const struct cmd_header_t *cmd_hdr;
     uint32_t cmd_buff_idx = 0;
     int32_t pop_ret_val;
-    int8_t status;
+    int32_t status;
 
     /* Create a shadow copy of data from SQ to L2 SCP */
     status = VQ_Prefetch_Buffer(vq_cached, vq_used_space, shared_mem_ptr, cmd_buff);
@@ -232,34 +231,32 @@ static inline void sqw_process_waiting_commands(uint32_t sqw_idx, vq_cb_t *vq_ca
     /* Update the tail offset in VQ shared memory so that host is able to push new commands */
     Host_Iface_Optimized_SQ_Update_Tail(vq_shared, vq_cached);
 
-    if(status != STATUS_SUCCESS)
+    if (status != STATUS_SUCCESS)
     {
-        Log_Write(LOG_LEVEL_ERROR, "SQW:ERROR:VQ prefetch buffer failed:%d\r\n",
-            status);
+        Log_Write(LOG_LEVEL_ERROR, "SQW:ERROR:VQ prefetch buffer failed:%d\r\n", status);
         return;
     }
 
     /* Process the commands buffer. */
-    while(cmd_buff_idx < vq_used_space)
+    while (cmd_buff_idx < vq_used_space)
     {
         /* Process commands from L2 SCP prefetched copy */
         pop_ret_val = VQ_Process_Command(cmd_buff, vq_used_space, cmd_buff_idx);
 
-        if(pop_ret_val > 0)
+        if (pop_ret_val > 0)
         {
             /* Set the command starting address */
-            cmd_hdr = (void*)&cmd_buff[cmd_buff_idx];
+            cmd_hdr = (void *)&cmd_buff[cmd_buff_idx];
 
-            Log_Write(LOG_LEVEL_DEBUG,
-                "SQW:Processing:SQW_IDX=%d:tag_id=%x:Popped_length:%d\r\n",
+            Log_Write(LOG_LEVEL_DEBUG, "SQW:Processing:SQW_IDX=%d:tag_id=%x:Popped_length:%d\r\n",
                 sqw_idx, cmd_hdr->cmd_hdr.tag_id, pop_ret_val);
 
             /* If barrier flag is set, wait until all cmds are
             processed in the current SQ */
-            if(cmd_hdr->cmd_hdr.flags & CMD_FLAGS_BARRIER_ENABLE)
+            if (cmd_hdr->cmd_hdr.flags & CMD_FLAGS_BARRIER_ENABLE)
             {
                 TRACE_LOG_CMD_STATUS(cmd_hdr->cmd_hdr.msg_id, (uint8_t)sqw_idx,
-                                     cmd_hdr->cmd_hdr.tag_id, CMD_STATUS_WAIT_BARRIER);
+                    cmd_hdr->cmd_hdr.tag_id, CMD_STATUS_WAIT_BARRIER);
 
                 sqw_command_barrier((uint8_t)sqw_idx);
             }
@@ -271,10 +268,9 @@ static inline void sqw_process_waiting_commands(uint32_t sqw_idx, vq_cb_t *vq_ca
             SQW_Increment_Command_Count((uint8_t)sqw_idx);
 
             /* Handle the command processing */
-            status = Host_Command_Handler(&cmd_buff[cmd_buff_idx],
-                (uint8_t)sqw_idx, start_cycles);
+            status = Host_Command_Handler(&cmd_buff[cmd_buff_idx], (uint8_t)sqw_idx, start_cycles);
 
-            if(status != STATUS_SUCCESS)
+            if (status != STATUS_SUCCESS)
             {
                 Log_Write(LOG_LEVEL_ERROR, "SQW:ERROR:Processing failed:%d\r\n", status);
             }
@@ -282,7 +278,7 @@ static inline void sqw_process_waiting_commands(uint32_t sqw_idx, vq_cb_t *vq_ca
             /* Update the command buffer index */
             cmd_buff_idx += (uint32_t)pop_ret_val;
         }
-        else if(pop_ret_val < 0)
+        else if (pop_ret_val < 0)
         {
             Log_Write(LOG_LEVEL_ERROR, "SQW:ERROR:VQ cmd processing failed:%d\r\n", pop_ret_val);
             SP_Iface_Report_Error(MM_RECOVERABLE, MM_SQ_PROCESSING_ERROR);
@@ -316,7 +312,7 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
 {
     uint64_t tail_prev;
     uint64_t start_cycles = 0;
-    void* shared_mem_ptr;
+    void *shared_mem_ptr;
     circ_buff_cb_t circ_buff_cached __attribute__((aligned(64)));
     vq_cb_t vq_cached;
 
@@ -325,24 +321,26 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
     vq_cb_t *vq_shared = Host_Iface_Get_VQ_Base_Addr(SQ, (uint8_t)sqw_idx);
 
     /* Make a copy of the VQ CB in cached DRAM to cached L1 stack variable */
-    ETSOC_Memory_Read_Local_Atomic((void*)vq_shared, (void*)&vq_cached, sizeof(vq_cached));
+    ETSOC_Memory_Read_Local_Atomic((void *)vq_shared, (void *)&vq_cached, sizeof(vq_cached));
 
     /* Make a copy of the Circular Buffer CB in shared SRAM to cached L1 stack variable */
-    ETSOC_Memory_Read_Uncacheable((void*)vq_cached.circbuff_cb, (void*)&circ_buff_cached,
-        sizeof(circ_buff_cached));
+    ETSOC_Memory_Read_Uncacheable(
+        (void *)vq_cached.circbuff_cb, (void *)&circ_buff_cached, sizeof(circ_buff_cached));
 
     /* Save the shared memory pointer */
-    shared_mem_ptr = (void*)vq_cached.circbuff_cb->buffer_ptr;
+    shared_mem_ptr = (void *)vq_cached.circbuff_cb->buffer_ptr;
 
     /* Verify that the head pointer in cached variable and shared SRAM are 8-byte aligned addresses */
-    if (!(IS_ALIGNED(&circ_buff_cached.head_offset, 8) && IS_ALIGNED(&vq_cached.circbuff_cb->head_offset, 8)))
+    if (!(IS_ALIGNED(&circ_buff_cached.head_offset, 8) &&
+            IS_ALIGNED(&vq_cached.circbuff_cb->head_offset, 8)))
     {
         Log_Write(LOG_LEVEL_ERROR, "SQW:SQ HEAD not 64-bit aligned\r\n");
         SP_Iface_Report_Error(MM_RECOVERABLE, MM_SQ_BUFFER_ALIGNMENT_ERROR);
     }
 
     /* Verify that the tail pointer in cached variable and shared SRAM are 8-byte aligned addresses */
-    if (!(IS_ALIGNED(&circ_buff_cached.tail_offset, 8) && IS_ALIGNED(&vq_cached.circbuff_cb->tail_offset, 8)))
+    if (!(IS_ALIGNED(&circ_buff_cached.tail_offset, 8) &&
+            IS_ALIGNED(&vq_cached.circbuff_cb->tail_offset, 8)))
     {
         Log_Write(LOG_LEVEL_ERROR, "SQW:SQ tail not 64-bit aligned\r\n");
         SP_Iface_Report_Error(MM_RECOVERABLE, MM_SQ_BUFFER_ALIGNMENT_ERROR);
@@ -353,7 +351,7 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
 
     Log_Write(LOG_LEVEL_INFO, "SQW:H%d:IDX=%d\r\n", hart_id, sqw_idx);
 
-    while(1)
+    while (1)
     {
         /* Update the SQW state to idle */
         atomic_store_local_32(&SQW_CB.sqw_status[sqw_idx].state, SQW_STATE_IDLE);
@@ -380,11 +378,11 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
         VQ_Get_Head_And_Tail(vq_shared, &vq_cached);
 
         /* Verify that the tail value read from memory is equal to previous tail value */
-        if(tail_prev != VQ_Get_Tail_Offset(&vq_cached))
+        if (tail_prev != VQ_Get_Tail_Offset(&vq_cached))
         {
             Log_Write(LOG_LEVEL_ERROR,
-            "SQW:FATAL_ERROR:Tail Mismatch:Cached: %ld, Shared Memory: %ld Using cached value as fallback mechanism\r\n",
-            tail_prev, VQ_Get_Tail_Offset(&vq_cached));
+                "SQW:FATAL_ERROR:Tail Mismatch:Cached: %ld, Shared Memory: %ld Using cached value as fallback mechanism\r\n",
+                tail_prev, VQ_Get_Tail_Offset(&vq_cached));
 
             SP_Iface_Report_Error(MM_RECOVERABLE, MM_SQ_PROCESSING_ERROR);
 
@@ -396,15 +394,15 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
         uint64_t vq_used_space = VQ_Get_Used_Space(&vq_cached, CIRCBUFF_FLAG_NO_READ);
 
         /* Process commands until there is no more data in VQ */
-        while(vq_used_space)
+        while (vq_used_space)
         {
             /* Process the pending commands */
-            sqw_process_waiting_commands(sqw_idx, &vq_cached, vq_shared, start_cycles,
-                shared_mem_ptr, vq_used_space);
+            sqw_process_waiting_commands(
+                sqw_idx, &vq_cached, vq_shared, start_cycles, shared_mem_ptr, vq_used_space);
 
             /* In case of SQW aborted state, we need to read the shared copy of head and tail
             again to make sure that all the commands in VQ are processed and aborted. */
-            if(atomic_load_local_32(&SQW_CB.sqw_status[sqw_idx].state) == SQW_STATE_ABORTED)
+            if (atomic_load_local_32(&SQW_CB.sqw_status[sqw_idx].state) == SQW_STATE_ABORTED)
             {
                 /* Refresh the cached VQ CB - Get updated head and tail values */
                 VQ_Get_Head_And_Tail(vq_shared, &vq_cached);
@@ -441,16 +439,17 @@ void SQW_Launch(uint32_t hart_id, uint32_t sqw_idx)
 void SQW_Decrement_Command_Count(uint8_t sqw_idx)
 {
     /* Decrement commands count being processed by current SQW */
-    int32_t original_val =
-        atomic_add_signed_local_32(&SQW_CB.sqw_status[sqw_idx].cmds_count, -1);
+    int32_t original_val = atomic_add_signed_local_32(&SQW_CB.sqw_status[sqw_idx].cmds_count, -1);
 
     if ((original_val - 1) < 0)
     {
-       Log_Write(LOG_LEVEL_ERROR, "SQW[%d] Decrement:Command Counter is Negative : %d\r\n", sqw_idx, original_val -1);
+        Log_Write(LOG_LEVEL_ERROR, "SQW[%d] Decrement:Command Counter is Negative : %d\r\n",
+            sqw_idx, original_val - 1);
     }
     else
     {
-       Log_Write(LOG_LEVEL_DEBUG, "SQW[%d] Decrement:Command Counter: %d\r\n", sqw_idx, original_val - 1);
+        Log_Write(LOG_LEVEL_DEBUG, "SQW[%d] Decrement:Command Counter: %d\r\n", sqw_idx,
+            original_val - 1);
     }
 }
 
@@ -477,10 +476,10 @@ void SQW_Decrement_Command_Count(uint8_t sqw_idx)
 void SQW_Increment_Command_Count(uint8_t sqw_idx)
 {
     /* Increment commands count being processed by current SQW */
-    int32_t original_val =
-        atomic_add_signed_local_32(&SQW_CB.sqw_status[sqw_idx].cmds_count, 1);
+    int32_t original_val = atomic_add_signed_local_32(&SQW_CB.sqw_status[sqw_idx].cmds_count, 1);
 
-    Log_Write(LOG_LEVEL_DEBUG, "SQW[%d] Increment:Command Count: %d\r\n", sqw_idx, original_val + 1);
+    Log_Write(
+        LOG_LEVEL_DEBUG, "SQW[%d] Increment:Command Count: %d\r\n", sqw_idx, original_val + 1);
 }
 
 /************************************************************************
@@ -512,8 +511,8 @@ void SQW_Abort_All_Pending_Commands(uint8_t sqw_idx)
     /* Traverse SQ and check for state. If busy, set abort state
     and wait for it to be idle. This would guarantee
     that all pending commadns in SQ are aborted. */
-    old_state = atomic_compare_and_exchange_local_32(&SQW_CB.sqw_status[sqw_idx].state,
-        SQW_STATE_BUSY, SQW_STATE_ABORTED);
+    old_state = atomic_compare_and_exchange_local_32(
+        &SQW_CB.sqw_status[sqw_idx].state, SQW_STATE_BUSY, SQW_STATE_ABORTED);
 
     if (old_state == SQW_STATE_BUSY)
     {
